Add files via upload
Future Academy Africa is an online institute that offers scholarships for students to learn Data Science. I enrolled in the Data Science program, which provided a foundational understanding of Machine Learning and Artificial Intelligence. The curriculum delved into the core concepts of Data Science, including the use of relevant libraries. The FAA.ipynb code encapsulates the essence of the learning experience.Future Academy Africa is an online institute that tutors students through scholarship and i opt-in for Data Science where basic knowledge about Machine Learning and Artificial Intelligence was basically about and dived thoroughly into what a Data Science is, libraries used and son on. The FAA.ipynb code says it all